Made sections of '/FlowDiagram' into multiple components. Attempt to fix hydration errors.
